+++ CHANGELOG.md
@@ -18,16 +18,20 @@ New:
 * #62 - Diagnostic for file extension `*.wapproj`, must match project type GUID
   * of `C7167F0D-BC9F-4E6E-AFE1-012C56B48DB5` (Windows Application Packaging)
 
-* Improvement:
+Improvement:
 
 * #63 - Support for unfinished project lines
 
-* Changes:
+Changes:
 
 * #61 - Project type `F# SDK-style` is no longer unknown
   * GUID `6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705`
 * #62 - Project type `Windows Application Packaging` is no longer unknown
   * GUID `C7167F0D-BC9F-4E6E-AFE1-012C56B48DB5`
+
+Fixes:
+
+* #58 - Diagnostic was not triggered when language was changed to `sln`
 
 ## [1.7.0]